deploy: ApplyParams: do not write params.env if there are no updates
Optimize file operations, do not rewrite params.env on every
reconcile if no updates for it. In practice updates happen only on
the first run.

Since updates happen in 2 loops, move actual update to own function
which returns "number of updates" (actually 0 or 1, so acts as
boolean) and use bitwise operations to accumulate updates for every
field.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>