logger: change visible level with zap settings
With logr it is possible to decrease logging priority of a subsystem
by passing logger.V(<verbosity>) down the stack, the <verbosity>
value is then added to the level of the message making it less
important.

This approach was used in the operator, but it causes some
inconsitences, like [1], where setting desired level to
production/error info messages are still shown.

Fix it by changing filtering on the logging backend level.

[1] https://issues.redhat.com/browse/RHOAIENG-12712?focusedId=25642670&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-25642670

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>