webhook: switch to contextual logger
Use k8s contextual logger instead of own. Comparing to other parts
of the operator webhook is a bit special since it serves own
endpoints and has context independent from controllers.

The output changes and it looks like

```
{"level":"info","ts":"2024-10-09T13:11:14Z","logger":"admission","msg":"Got request","object":{"name":"default-dsci"},"namespace":"","name":"default-dsci","resource":{"group":"dscinitialization.opendatahub.io","version":"v1","resource":"dscinitializations"},"user":"kube:admin","requestID":"e3489544-38dc-410a-8c96-56d6f9f90cd6","operation":"CREATE","kind":"DSCInitialization"}
```

(the message is not from the current codebase, was added for demo only)

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>